app/src/main/java/com/techcourse/service/TxUserService.java
package com.techcourse.service;

import com.interface21.transaction.support.TransactionManager;
import com.techcourse.domain.User;

public class TxUserService implements UserService {

    private final UserService userService;
    private final TransactionManager transactionManager;

    public TxUserService(UserService userService, TransactionManager transactionManager) {
        this.userService = userService;
        this.transactionManager = transactionManager;
    }

    @Override
    public User findById(long id) {
        return transactionManager.performTransaction(() -> userService.findById(id));
    }

    @Override
    public void save(User user) {
        transactionManager.performTransaction(() -> userService.save(user));
    }

    @Override
    public void changePassword(final long id, final String newPassword, final String createdBy) {
        transactionManager.performTransaction(() -> userService.changePassword(id, newPassword, createdBy));
    }
}
